Created new gameLogic.js script to futher separate concerns from deal.js script
