+++ scripts/deal.js
@@ -1,6 +1,5 @@
 // Imports the variables below from base.js
 import { deck, discardPile, playerPlay, computerPlay } from "./base.js";
-import { lose, refresh } from "./playerMoney.js";
 
 // Initialize variables that track card values as points
 let playerTotal = 0;
@@ -94,16 +93,10 @@ function createComputerCard() {
                 break;
         }
     }
-
-    // // Ensures that the computer's first card is face-down instead of face-up
-    // const dealerFirstCardImg = document.getElementById('computerCards').getElementsByClassName('cardImg');
-    // dealerFirstCardImg[0].src = 'images/card_back.png';
-    // const dealerFirstCardText = document.getElementById('computerCards').getElementsByClassName('cardText');
-    // dealerFirstCardText[0].style.visibility = 'hidden';
 }
 
+// Ensures that the computer's first card is face-down instead of face-up
 function faceDown() {
-    // Ensures that the computer's first card is face-down instead of face-up
     const dealerFirstCardImg = document.getElementById('computerCards').getElementsByClassName('cardImg');
     dealerFirstCardImg[0].src = 'images/card_back.png';
     const dealerFirstCardText = document.getElementById('computerCards').getElementsByClassName('cardText');
@@ -156,129 +149,12 @@ function start() {
     }
 }
 
-// Disables actions buttons except Clear when the player goes over 21
-function bust() {
-    const actionButtons = document.getElementById('actions').getElementsByClassName('actionButton');
-    for (let i = 0; i < actionButtons.length - 1; i++) {
-        actionButtons[i].disabled = true;
-    }
-    lose(0);
-    refresh();
-}
-
-// Compares your card and the computer's card to see which card has a higher value
-function compare() {
-    // Sets the value of the computer's face cards and aces to a number; otherwise, converts the cards' value from a string to a number
-    for (let i = 0; i < computerPlay.length; i++) {
-        switch (computerPlay[i].Value) {
-            case ('J'):
-            case ('Q'):
-            case ('K'):
-                computerPlay[i].Points = 10;
-                computerPoints.push(computerPlay[i].Points);
-                break;
-            case ('A'):
-                if (computerTotal <= 10) {
-                    computerPlay[i].Points = 11;
-                    computerPoints.push(computerPlay[i].Points);
-                } else {
-                    computerPlay[i].Points = 1;
-                    computerPoints.push(computerPlay[i].Points);
-                }
-                break;
-            default:
-                computerPlay[i].Points = Number(computerPlay[i].Value);
-                computerPoints.push(computerPlay[i].Points);
-                break;
-        }
-        // Adds all of the points from the computer's cards to computerTotal
-        computerTotal = computerPoints.reduce((a, b) => a + b, 0);
-    }
-    
-    // Sets the value of the player's face cards and aces to a number; otherwise, converts the cards' value from a string to a number
-    for (let i = 0; i < playerPlay.length; i++) {
-        switch (playerPlay[i].Value) {
-            case ('J'):
-            case ('Q'):
-            case ('K'):
-                playerPlay[i].Points = 10;
-                playerPoints.push(playerPlay[i].Points);
-                break;
-            case ('A'):
-                if (playerTotal <= 10) {
-                    playerPlay[i].Points = 11;
-                    playerPoints.push(playerPlay[i].Points);
-                } else {
-                    playerPlay[i].Points = 1;
-                    playerPoints.push(playerPlay[i].Points);
-                }
-                break;
-            default:
-                playerPlay[i].Points = Number(playerPlay[i].Value);
-                playerPoints.push(playerPlay[i].Points);
-                break;
-        }
-        // Adds all of the points from the player's cards to playerTotal
-        playerTotal = playerPoints.reduce((a, b) => a + b, 0);
-    }
-    console.log(playerPoints);
-    console.log(playerTotal);
-
-    // Shows the player the value of the cards in their hand
-    document.getElementById('playerTotal').textContent = `Player Total: ${playerTotal}`;
-}
-
-function checkForBust() {
-    // Initialize variables that track whether the player has an ace and if an ace has had its value reduced to 1
-    const lookForAce = playerPlay.find(item => item.Value === 'A');
-    let aceAltered = false;
-    
-    // Reduces an ace's value from 11 to 1 if it would cause the player to go over 21
-    if (playerTotal > 21 && lookForAce != undefined) {
-        if (aceAltered = false) {
-            for (let i = 0; i < playerPlay.length; i++) {
-                if (playerPlay[i].Value === 'A') {
-                    playerPlay[i].Points = 1;
-                    playerPoints[i] = 1;
-                    playerTotal = playerPoints.reduce((a, b) => a + b, 0);
-                    aceAltered = true;
-                    console.log(playerPoints);
-                    console.log(playerTotal);
-                }
-            }
-            console.log(lookForAce);
-            console.log(aceAltered);
-        } else if (aceAltered = true) {
-            bust();
-            console.log("Busted!");
-        }
-        
-    } else if (playerTotal > 21 && lookForAce === undefined) {
-        bust();
-        console.log("Busted!");
-    }
-
-    // Reduces an ace's value from 11 to 1 if it would cause the computer to go over 21
-    if (computerTotal > 21) {
-        for (let i = 0; i < computerPlay.length; i++) {
-            if (computerPlay[i].Value === 'A') {
-                computerPlay[i].Points = 1;
-                playerPoints[i] = 1;
-                computerTotal = playerPoints.reduce((a, b) => a + b, 0);
-                console.log(playerPoints);
-                console.log(computerTotal);
-            }
-        }
-    }
-
-    // Shows the player the value of the cards in their hand
-    document.getElementById('playerTotal').textContent = `Player Total: ${playerTotal}`;
-}
-
-// Clears the contents of the playerPoints array
+// Resets the values of the below variables
 function clearPoints() {
     playerPoints = [];
     computerPoints = [];
+    playerTotal = 0;
+    computerTotal = 0;
 }
 
 // Allows other scripts to change playerTotal
@@ -292,4 +168,4 @@ function increasecomputerTotal(amount) {
 }
 
 // Exports all functions and variables to index.js and other scripts that need these
-export { playerPoints, playerTotal, computerPoints, computerTotal, createPlayerCard, createComputerCard, faceDown, deal, start, compare, checkForBust, increasePlayerTotal, increasecomputerTotal, clearPoints };
+export { playerPoints, playerTotal, computerPoints, computerTotal, createPlayerCard, createComputerCard, faceDown, deal, start, clearPoints, increasePlayerTotal, increasecomputerTotal };
